Trait fallible_collections::arc::FallibleArc

pub trait FallibleArc<T> {
    fn try_new(t: T) -> Result<Self, TryReserveError>
    where
        Self: Sized;
}

👎 Deprecated since 0.3.1:
⚠️️️this function is not completely fallible, it can panic !, see issue. help wanted

trait to implement Fallible Arc

Required methods

fn try_new(t: T) -> Result<Self, TryReserveError> where
Self: Sized,

👎 Deprecated since 0.3.1:
⚠️️️this function is not completely fallible, it can panic !, see issue. help wanted

try creating a new Arc, returning a Result<Box, TryReserveError> if allocation failed

Loading content...