Trait fallible_collections::vec::FallibleVec

pub trait FallibleVec<T> {
    fn try_reserve(&mut self, additional: usize) -> Result<(), TryReserveError>;

    fn try_push(&mut self, elem: T) -> Result<(), TryReserveError>;

    fn try_push_give_back(
        &mut self, 
        elem: T
    ) -> Result<(), (T, TryReserveError)>;

    fn try_with_capacity(capacity: usize) -> Result<Self, TryReserveError>
    where
        Self: Sized;

    fn try_insert(
        &mut self, 
        index: usize, 
        element: T
    ) -> Result<(), (T, TryReserveError)>;

    fn try_append(&mut self, other: &mut Self) -> Result<(), TryReserveError>;

    fn try_resize(
        &mut self, 
        new_len: usize, 
        value: T
    ) -> Result<(), TryReserveError>
    where
        T: Copy + Clone;

    fn try_resize_with<F>(
        &mut self, 
        new_len: usize, 
        f: F
    ) -> Result<(), TryReserveError>
    where
        F: FnMut() -> T;

    fn try_resize_no_copy(
        &mut self, 
        new_len: usize, 
        value: T
    ) -> Result<(), TryReserveError>
    where
        T: TryClone;

    fn try_extend_from_slice(
        &mut self, 
        other: &[T]
    ) -> Result<(), TryReserveError>
    where
        T: Copy + Clone;

    fn try_extend_from_slice_no_copy(
        &mut self, 
        other: &[T]
    ) -> Result<(), TryReserveError>
    where
        T: TryClone;
}

trait implementing all fallible methods on vec

Required methods

fn try_reserve(&mut self, additional: usize) -> Result<(), TryReserveError>

see reserve

fn try_push(&mut self, elem: T) -> Result<(), TryReserveError>

see push

fn try_push_give_back(&mut self, elem: T) -> Result<(), (T, TryReserveError)>

try push and give back ownership in case of error

fn try_with_capacity(capacity: usize) -> Result<Self, TryReserveError> where
Self: Sized,

see with capacity, (Self must be sized by the constraint of Result)

fn try_insert(
&mut self,
index: usize,
element: T
) -> Result<(), (T, TryReserveError)>

see insert

fn try_append(&mut self, other: &mut Self) -> Result<(), TryReserveError>

see append

fn try_resize(
&mut self,
new_len: usize,
value: T
) -> Result<(), TryReserveError> where
T: Copy + Clone,

see resize, only works when the value implements Copy, otherwise, look at try_resize_no_copy

fn try_resize_with<F>(
&mut self,
new_len: usize,
f: F
) -> Result<(), TryReserveError> where
F: FnMut() -> T,

fn try_resize_no_copy(
&mut self,
new_len: usize,
value: T
) -> Result<(), TryReserveError> where
T: TryClone,

resize the vec by trying to clone the value repeatingly

fn try_extend_from_slice(&mut self, other: &[T]) -> Result<(), TryReserveError> where
T: Copy + Clone,

see resize, only works when the value implements Copy, otherwise, look at try_extend_from_slice_no_copy

fn try_extend_from_slice_no_copy(
&mut self,
other: &[T]
) -> Result<(), TryReserveError> where
T: TryClone,

extend the vec by trying to clone the value in other

Loading content...